RTS-7157 User sebra sync: up the number of treated users
Some users have been unexpectedly blocked ad/or not synchronised.
The reasons seem to be that (1) these user have been marked as not
visible persons in Sebra (but visible as staff on area), or (2)
for other reasons not current in Sebra, but added manually as
content managers or staff in w3.

Since these are not listed in the Sebra query "persons?type=",
the solution was to get those users separately and add them at
the end of the list. Thus, they get synchronised and properly
blocked or unblocked.

The synchronisation script was also speed optimised by
switching to using dbquery() and only performing user_load()
when a change is really needed.