RTS-7161 CSS arrowhead added to ui-tabs-selected
CSS arrowhead added to enhance the understanding of clickable
elements on pages with the ui-tabs-selected nodes

The reason this works is simple enough. The idea is a box with
zero width and height. The actual width and height of the arrow is
determined by the width of the border. In this particular situation
the right border is transparent, the top and bottom border are
colored, leaving an arrowhead inward shape.