RTS-6434 CRIStin integration for user profile page
Implements the display of publications directly from CRIStin on the
user profile pages.

It expands the uib_cristin module, and fetches, sorts, and formats
publication data from the CRIStin web service (via json output).

Publications are displayed in two places: Selected publications,
and in the Publications tab. The following applies to the latter:

  The number of main categories fetched are the four most relevant
  as scientific publications (out of 12 in CRIStin): BOK,
  TIDSSKRIFTPUBL, RAPPORT, BOKRAPPORTDEL.

  The user may edit these settings: The maximum number of publica-
  tions to display, sort order (category or date), placement of the
  publications from CRIStin in relation to manually added ones.