RTS-7507 Display not only the link for the area but also the title
In RTS-7030 we wanted to add a link to the ou an user was affiliated
to. The OUs should not be targeted so the link displayed is the url of
the area connected to the OU. The link text however remained the title
of the OU. As the OU needs to be translated we have decided to display
the title of the area as well.