RTS-7161 Moved the sass into the offices module and made the code DRYer
The code was initially inserted into nodes.scss but it really belonged
in the offices module. I also made the arrow styling into a mixin to
be reused in other cases.

A little clean up of office tabs styling were needed due to the fact
that some office related styling had been inserted into the banner
field and made overrides for our css.