RTS-7617 Avoid type casting
It seems that type casting and array to string did not work as
intended on the staging server (running php 5.3) but locally
(running php 5.5). The type casting was necessary because we
would get an array instead of a string from CRIStin in certain
cases - in particular when there was no data.

This commit instead checks if we get an array, and collapsing it
before feeding the resulting string to check_plain().

The commit also adds a check_plain for the author list, which had
been left out before.