RTS-7612 Manipulate user choices when adding content to menu
To prevent users adding more than five top level menu items we want
to limit their choices when adding content to area menus. By removing
choice to insert a menu item as a sibling to top level menu items and
make the relationship with selected menu item a required value we
achieve this.