RTS-7659 Corrected styling after RTS-6331 rebase
Small changes in uib_ansatte.view_default to make the previously
added styling work properly. No longer excluding field
field_uib_user_competence, and removed the HTML code to add
a list of fields of competence in the info div for persons list.

Added a specific styling for the a.active link to compensate for
the fact that styling coming from the navigation css file were
with wrong color.