RTS-7669 Do not show CRIStin link when there are no publications
The addition of the CRIStin integration had inadvertently caused
the publications data to never be empty. This commit makes sure
nothing is output in the publications tab when there are no
publications.

However, if there are publications present, but the user has set
the maximum number of publications to display to zero, there will
be a link to CRIStin.