RTS-7830 Ignore Entity Translation fallback for fields on user
Certain fields on the user entity is setup with Entity Translation
and we have set a fallback that shows another language if the
translation for the current language is missing. This fallback may
be important in content types, but is not desired on the user
profile page. However, the setting is global.

We solve this by introducing a new field formatter that ignores
the fallback setting and only displays content if it exists in the
current language. The formatter works for text, text_long, link,
and taxonomy_term_reference link.

The address field was already dealt with in uib_user_user_view_alter()

The uib_cristin module had to be slightly modified in order to not
mess up the tabs on the user page.