RTS-7848 Trim incoming phone numbers during Sebra sync
During Sebra synchronisation, any content would be considered as
presence of a phone number (and prefixing it with "+47").

It turns out that user phone numbers from Sebra may consist of only
whitespace, and this was not trimmed properly, thus creating bogus
phone numbers containing a list of "+47 ".

Fixed by cleaning input from Sebra.