RTS-7978 Make sure publications field title is translated
When the user has not added any own content to the publications
field, we create the field render array (to put CRIStin data
in it). The title needs to be wrapped in the t() function in
order to be translated and displayed correcty in the tabs on
the user profile page.