RTS-7962 Remove user aliases from non-main users
A person can have more than one active user account, but one is
the main one. The aliases of these acounts may interfere.

The solution is to remove aliases for the user that is not the main
user. The commit introduces a script and sets it up to run in by
bin/post-upgrade.