RTS-7902 Added view with search and filter for files
The view added have the option to search for file names and have
the option to filter on document type. In addition there is added
sorting on the list, giving the option to sort on various fields.

The sorting files are: Name, filesize, title, type, upload date
and use count. It's unknown if this is a function that will be
utilized, but it's now an option.