RTS-8018 Reordered user information in staff listings
The order of user information had been altered causing the user
competence field to be displayed above the the rest of the user
information at small screen sizes.

This commit reorders the user information in the staff listing view
and adds styling to display the user information correct on small
screens.