RTS-7995 Add user slug and alias creation to sebra sync
Adds slug creation. The user slug is created based on the user's
first and last names with a dot in between. If the slug is in use,
a suffix is added: an underscore plus a serial number, such as:
firstname.lastname_2

If the user overrides names in the user profile, the slug will
change and a new alias will be added.