RTS-8145 Made selector target the correct sublist in study listings
The list of specializations in the first column of a study programme
had a markup that made the selector targeting the specializations in
a study listing view kick in. To avoid this behaviour we make sure
the selector targets a list of specializations only when presented in
the context of a views page.