RTS-8132 CRIStin data: Supervisors should not be authors
Supervisors for theses and similar reports are listed like the
authors in the data from CRIStin, but they should not be presented
as authors. Nor should those theses be listed as publications in
the publication tab of the supervisor.

To solve this, we obtain the CRIStin person id, and check every
person connected to the report/thesis. The person's role in relation
to the publication is nested deeply into the structure, and the
relevant role is "VEILEDER". We do this early, right after the
publications are fetched from CRIStin, in order to get the publica-
tion count and formatting correct.