RTS-7850 Removed empty ul tag
When building the user profile an ul tag is added to contain the
tabs displaying data entered by users. When no data is entered in the
relevant fields an empty ul tag is found in the markup. By adding a
test to verify that content for the tabs exists we avoid polluting our
markup.