RTS-8080 Translated article shows in both Norwegian and English area
Current solution based on previous solution that assumed that if the
area exists its translation exists too.

New functionality:
Before: The id of the area in position 0 was replaced for the translated
  one. The new language assigned to the node.
Now: The area in position 0 is treated the same way as before. Areas in
  position 1 and higher are translated.

Refactoring of already existing functionality:
Before: Language of node changed on every submit. Then changed again
  if submit was of type translate.
Now: Language change happens only once. (outer if checks if submit is
  of type translate)