RTS-8158 Add Linked-in as following possiblility
Adds linkedin to the field formatter as a possibility. The icon had
already been added previously.

Linked-in uses different urls when viewing personal accounts and
schools. Because of this, the defaults in the list of recognised
social media was modified to include the url.

There is an account number in linked-in which may be used to access
the personal profile: linkedin.com/profile/view?id=11111111 as well
as a personal url: linkedin.com/in/personal-url.

Since the id is somewhat obsure, and seems to be available only by
copying it from the browser url, I have used the personal url as
identifier.