RTS-4377 "find content" should list external content as type
Solves the problem described in comment #21. An other consequence of
this solution is that the values of the options of the new dropdown box
are now numbers instead of strings. If the order of the options is
changed then so is their corresponding value. If the option "Event" is
moved then the code in uib_webdesk.module that controls when the event
type dropdown box is enabled should be updated. To mitigate this risk
event has been changed to the first option of the list.